requirements.txt
tensorflow>=2.10
pandas
numpy
scikit-learn
Pillow
joblib
